+++ tests/spend_test.rs
@@ -17,7 +17,7 @@ mod spend_tests {
         let params = TransactionParams {
             address: "tb1pspfcrvz538vvj9f9gfkd85nu5ty98zw9y5e302kha6zurv6vg07s8z7a8w".to_string(),
             amount: 2003,
-            fee_rate: 1,
+            fee_rate: 1.0,
             selected_outputs: vec![],
             note: Some("not a note".to_string()),
             tag: Some("hello".to_string()),
@@ -36,7 +36,7 @@ mod spend_tests {
         let params = TransactionParams {
             address: "tb1pspfcrvz538vvj9f9gfkd85nu5ty98zw9y5e302kha6zurv6vg07s8z7a8w".to_string(),
             amount: 4000,
-            fee_rate: 2,
+            fee_rate: 2.0,
             selected_outputs: vec![],
             note: Some("not a note".to_string()),
             tag: Some("hello".to_string()),
@@ -55,7 +55,7 @@ mod spend_tests {
         let params = TransactionParams {
             address: "tb1pspfcrvz538vvj9f9gfkd85nu5ty98zw9y5e302kha6zurv6vg07s8z7a8w".to_string(),
             amount: 1000,
-            fee_rate: 2,
+            fee_rate: 2.0,
             selected_outputs: vec![],
             note: Some("not a note".to_string()),
             tag: Some("hello".to_string()),
@@ -93,7 +93,7 @@ mod spend_tests {
         let params = TransactionParams {
             address: address.to_string(),
             amount: 1000,
-            fee_rate: 2,
+            fee_rate: 2.0,
             selected_outputs: vec![],
             note: Some("not a note".to_string()),
             tag: Some("hello".to_string()),
@@ -116,7 +116,7 @@ mod spend_tests {
         // verify transaction properties
         assert_eq!(transaction.amount, -1000);
         assert_eq!(transaction.address, address);
-        assert_eq!(transaction.fee_rate, 2);
+        assert_eq!(transaction.fee_rate, 2u64);
         assert_eq!(transaction.note, params.note);
         assert_eq!(transaction.get_change_tag(), params.tag);
     }
@@ -145,7 +145,7 @@ mod spend_tests {
         let transaction = draft_transaction.transaction.clone();
         assert_eq!(transaction.address, params.address);
         assert_eq!(transaction.amount, -(params.amount as i64));
-        assert_eq!(transaction.fee_rate, params.fee_rate);
+        assert_eq!(transaction.fee_rate, params.fee_rate as u64);
         assert_eq!(transaction.note, params.note);
         assert_eq!(transaction.get_change_tag(), params.tag);
     }
